% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{between_condition_correct}
\alias{between_condition_correct}
\title{Get between condition stats for correctness data}
\usage{
between_condition_correct(data, digits = 4)
}
\description{
Get between condition stats for correctness data
}
