% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{between_condition_flip}
\alias{between_condition_flip}
\title{Get between condition stats for flip data}
\usage{
between_condition_flip(data, digits = 4)
}
\description{
Get between condition stats for flip data
}
