% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{between_condition_helpful}
\alias{between_condition_helpful}
\title{Get between condition stats for helpfulness data}
\usage{
between_condition_helpful(data, digits = 4)
}
\description{
Get between condition stats for helpfulness data
}
