% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{between_condition_response}
\alias{between_condition_response}
\title{Get between conditions stats for first response data}
\usage{
between_condition_response(data, digits = 4)
}
\description{
Get between conditions stats for first response data
}
