% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{between_condition_stats}
\alias{between_condition_stats}
\title{Returns a tibble summarizing the results of a between conditions comparison
with a Fisher's exact test}
\usage{
between_condition_stats(data, variable, variable_option, digits = 4)
}
\arguments{
\item{data}{Formatted data from exp 1, 2, or 3}

\item{variable}{The variable you want to compare values of}

\item{variable_option}{The value of the variable you want the results in terms of}

\item{digits}{Number of significant digits to round the p-value to (default is 4)}
}
\description{
Returns a tibble summarizing the results of a between conditions comparison
with a Fisher's exact test
}
