% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting_funs.R
\name{format_data}
\alias{format_data}
\title{Format choice data}
\usage{
format_data(data)
}
\arguments{
\item{data}{A tibble consisting of the raw data from exp 1, 2, or 3}
}
\value{
A tibble
}
\description{
Format choice data
}
