% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{format_for_between_stats}
\alias{format_for_between_stats}
\title{Create a tibble with the number and percentage of children in each condition
with variable = variable_option}
\usage{
format_for_between_stats(data, variable, variable_option)
}
\arguments{
\item{data}{Formatted data from exp 1, 2, or 3}

\item{variable}{Unquoted name of the variable you want to look at}

\item{variable_option}{String representing the value of the variable for which
you want the number/percentage of children}
}
\value{
A tibble with the number of children with "variable_option" for variable
per condition (num_bb and num_bt) and the percentage of children with "variable_option"
for variable per condition (bb_"variable_option" and bt_"variable_option")
}
\description{
Create a tibble with the number and percentage of children in each condition
with variable = variable_option
}
