% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_stats_funs.R
\name{format_for_within_stats}
\alias{format_for_within_stats}
\title{Helper function that formats data for binomial tests}
\usage{
format_for_within_stats(data, variable, variable_option)
}
\arguments{
\item{data}{formatted data from exp 1, 2, or 3}

\item{variable}{variable to test}

\item{variable_option}{value of variable to test in terms of}
}
\description{
Helper function that formats data for binomial tests
}
