% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting_funs.R
\name{format_sr_data}
\alias{format_sr_data}
\title{Format social referencing data}
\usage{
format_sr_data(data, original_data)
}
\arguments{
\item{data}{A tibble consisting of the raw social referencing data from exp 1, 2, or 3}
}
\value{
A tibble
}
\description{
Format social referencing data
}
