% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_stats_funs.R
\name{get_binom_p_value}
\alias{get_binom_p_value}
\title{Returns p-value from a binomial test}
\usage{
get_binom_p_value(k, n, digits = 4, alternative = "two.sided")
}
\arguments{
\item{k}{number of successes}

\item{n}{total number}

\item{digits}{number of significant digits of p-value}

\item{alternative}{to pass to binom.test(); "two.sided", "greater", or "less"}
}
\description{
Returns p-value from a binomial test
}
