% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_stats_funs.R
\name{get_fishers_p_value}
\alias{get_fishers_p_value}
\title{Get a p-value from a Fisher's exact test}
\usage{
get_fishers_p_value(table, digits = 4)
}
\arguments{
\item{table}{A table}
}
\description{
Get a p-value from a Fisher's exact test
}
