% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat_funs.R
\name{get_num_excluded}
\alias{get_num_excluded}
\title{Create a tibble with counts of excluded kids for different reasons}
\usage{
get_num_excluded(data, exclude_code_var, ...)
}
\arguments{
\item{data}{Raw data from exp 1, 2, or 3}

\item{exclude_code_var}{Unquoted variable name of the variable specifying the
reason for exclusion}

\item{...}{Unquoted names of the exclude codes that you want in the resulting
tibble. If you want to include all of them, set this equal to \code{everything()}}
}
\value{
A tibble whose columns are the specified exclusion codes and values are
the number of children excluded under that code. The \code{total} column sums
over all specified exclusion codes (not the total number of excluded kids for
any reason)
}
\description{
Create a tibble with counts of excluded kids for different reasons
}
\examples{
get_num_excluded(three_toy, excludeCode, parent, sibling, exp)
}
