% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat_funs.R
\name{get_percentage_female}
\alias{get_percentage_female}
\title{Calculate the percentage of female participants}
\usage{
get_percentage_female(data, ..., num_digits = 0)
}
\arguments{
\item{data}{A tibble (formatted data from exp 1, 2, or 3)}

\item{...}{Any criteria you want to filter by}

\item{num_digits}{The number of significant digits}
}
\description{
Calculate the percentage of female participants
}
\examples{
get_percentage_female(three_toy_tidy, 3, condition == "Broken Button")
}
