% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat_funs.R
\name{get_summ_stat}
\alias{get_summ_stat}
\title{Calculate the mean or standard deviation of a variable in a tibble}
\usage{
get_summ_stat(data, summary_fun, summary_var, ..., num_digits = 2)
}
\arguments{
\item{data}{A tibble (formatted data from exp 1, 2, or 3)}

\item{summary_fun}{The summary function you want to apply}

\item{summary_var}{The variable you want to summarise}

\item{...}{Any criteria you want to filter by}

\item{num_digits}{The number of significant digits you want to round the
result of summary_fun() to. Default is 2}
}
\description{
Calculate the mean or standard deviation of a variable in a tibble
}
\examples{
get_summ_stat(three_toy_tidy, mean, age, condition == "Broken Button", num_digits = 3)
}
