% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat_funs.R
\name{summary_tibble}
\alias{summary_tibble}
\title{Create a tibble various summary statistics (mean and sd ages; percentage female;
number of participants; number per condition; number per location)}
\usage{
summary_tibble(data)
}
\arguments{
\item{data}{A tibble (formatted data from exp 1, 2, or 3)}
}
\description{
Create a tibble various summary statistics (mean and sd ages; percentage female;
number of participants; number per condition; number per location)
}
